[msbuild] introduce `_CompileAppManifestInputs` (#22745)
Fixes: https://github.com/dotnet/macios/issues/19607

The `_CompileAppManifest` MSBuild target currently runs on every build
because it has no `Inputs` defined.

In cases where the inputs are not discretely known, we can use a pattern
we've used in dotnet/android:

* https://github.com/dotnet/android/blob/b61cac33ea1a1a8c4751b8f9f1eade8cf27313c6/src/Xamarin.Android.Build.Tasks/Xamarin.Android.Common.targets#L960-L990

* Write a file with `<WriteLinesToFile WriteOnlyWhenDifferent="true">`
to the `obj` directory with all the parameters to the
`<CompileAppManifest/>` MSBuild task.

* Use this file as the `Inputs` for the `_CompileAppManifest` target.

* If any of the parameters change, the file will be different and
`_CompileAppManifest` will run again.

Updated `tests/dotnet/UnitTests/WindowsTest.cs` and other tests to verify this target
is skipped appropriately.

The `<CompileAppManifest/>` MSBuild task also no longer has to do any "up-to-date"
checks of its own.

Co-authored-by: Rolf Bjarne Kvinge <rolf@xamarin.com>