[vs] Disable build acceleration (#23332)
Visual Studio Build Acceleration is a feature designed to avoid rebuilding a project when changes occur in its project references.

For example, if you have an iOS app project and a MAUI class library that includes XAML files, an incremental change to the XAML will not trigger a rebuild of the iOS project. Instead, the updated assembly is simply copied into the iOS project’s output directory.

However, this behavior is not compatible with iOS builds, since the final app bundle is produced as part of the build process. As a result, skipping the build step can lead to unexpected behavior as incremental changes not reflected in the app.

For more details, see the https://github.com/dotnet/project-system/blob/main/docs/build-acceleration.md.

---------

Co-authored-by: Emanuel Fernandez Dell'Oca <3286258+emaf@users.noreply.github.com>