[RGen] Add support for the generation of constructors. (#23341)
This commit adds support to generate the constructors of a class. This
comes with some extra changes to take a look at:

1. Constructor formatter: It now takes an extra parameter that can be
used to state the addition of calling base (NSObjectFlag.Empty).
2. Fixed the messsage send generation to use the this.SuperHandle. We
had a bug and we were generating the super call with the this.Handle.
3. Update all tests according to the changes for the super calls.
4. Added the constructor generation in the class emitter.

We still have to correctly add the Skip method for the constructors that
are not partial. This will need to be added in the net10.0 branch in
which we will have partial constructors.

---------

Co-authored-by: GitHub Actions Autoformatter <github-actions-autoformatter@xamarin.com>