[Windows] Simplify LoadLinkerOutput (#23473)
This aims to improve reading the linker output files remotely from
Windows by executing the `ReadItemsFromFile` task once (and having a
single network roundtrip). This is a minor improvent as in a fast
connection this improves in around ~150ms (~340ms to ~70ms), but in a
slow one it could make a big difference.

Fixes https://github.com/dotnet/macios/issues/19601