[Windows] Make LinkNativeCode create output file
The `_LinkNativeExecutable` always runs because its inputs/outputs are not created on Windows. Its own output file was never created because the task was not configured to do so.

Also, the `_MtouchSymbolsList` was being written remotely so it was never found on Windows when checking the outputs. It's safe to write this locally as it will be copied as part of running this task as it is one of its declared dependencies.

Fixes https://github.com/dotnet/macios/issues/19610