[Windows] Add inputs/outputs to _CreatePkgInfo (#23553)
Makes the `CreatePkgInfo` task to produce an output file on Windows, and
add the same as input of the `_CreatePkgInfo` target so if won't run
again while the file exists as it is a static file.

This is a minor improvement but affects every single incremental build.

Fixes https://github.com/dotnet/macios/issues/19608.