[devops] Fix hiding comments. (#23557)
Some comments fail to hide because:

> Warning: Failed to hide comment for non-latest commit: Cannot convert argument "restCommentId", with value: "3177717194", for "HideNewlyPostedComment" to type "System.Int32": "Cannot convert value "3177717194" to type "System.Int32". Error: "Value was either too large or too small for an Int32.""

Looking at the code, I found an area of improvement in HideNewlyPostedComment:
we have the underlying id for the comment we want to hide, yet we search all
the comments in the PR to find those matching a certain string and hide those.
Simplify this to just hide the comment with the underlying id. Additionally
pass the underlying id around as a string, that way we avoid any problems with
parsing integers. This made the entire HideNewlyPostedComment unused, so
remove it.