[RGen] Move the BindingTypeSemanticAnalyzer to use the Binding struct and a set of validators. (#23623)
Move the analyzer to use the Validator approach, this removes the need
to add some code duplication to parse the AST. This new approach re-uses
the data model of rgen to parse the AST and uses a validator to ensure
that the data struct is correct.

The SmartEnumsAnalyzer will be remove in later commits as soon as we
have the SmartEnumValidator implemented with the same checks.