[RGen] Handle sealed property and methods correctly.
From the [review](https://github.com/dotnet/macios/pull/23656#pullrequestreview-3141249358) we noticed
that sealed property and methods do not need the IsDirecBinding check.
We update the struct to let use know if the method is sealed and
generate the call accordingly. We just need to update some if
conditions.

Tests have been updated to handle this case.