Add a new flag to skip the registration.
The idea is to use a flag to let the registrar know that a exported
symbol can be skipped. Because we want not to change the registrar at
this moment, we add the old Export attribute in methods, properties and
constructors for the registrar to parse. If the SkipRegistration flag is
set, we do not add the old bgen attribute which means that it will be
ignored by the registrar.