[RGen] Raise a warning in an async method that will use a nameless tuple. (#23776)
If a method is marked as async and does not provide a type or a result
type name the generator will use a nameless tuple. The nameless tuple
API is not very nice (Item1, Item2 etc). The analyzer will raise a
warning in this situation so that the developer creates a better API.

---------

Co-authored-by: GitHub Actions Autoformatter <github-actions-autoformatter@xamarin.com>