[RGen] Add a new type to track versions. (#23838)
When adding a platform to a symbol it might be the case that the
addition has an implication for another platform. For example, when we
add support for iOS we are adding support for Catalyst. This new struct
keeps track of the reason why a version was added and will allow to
choose always the Explicit version when two versions are compared.

This will later can be used with the Builder class to add Implicit
versions for a platform and be able to calculate which version to use.

---------

Co-authored-by: GitHub Actions Autoformatter <github-actions-autoformatter@xamarin.com>