[msbuild] Improve detection of resources that would end up outside the app bundle. (#23919)
Due to #23898, the build may end up with resources whose LogicalName property
(which is a path relative to the app bundle where the resource in question
would be located inside the app bundle) points to a location outside the app
bundle.

On macOS we correctly detect this and report an error:

> The path '../../../../../../../../../Users/rolf/test/bugs/issue2576178/RazorClassLibrary/wwwroot/background.png' would result in a file outside of the app bundle and cannot be used.

But no such thing was reported on Windows.

Additionally, this become a problem in .NET 10 on Windows because we now
process resources in library projects on Windows without a connection to a
Mac, while previously a connection to a Mac would be required to run into this
problem.

So:

* Fix the check to work on Windows as well.
* Also make it a warning (and just ignore the file in question) instead of an
  error, to keep the "works-by-luck" behavior of .NET 9.

This works around #23898 while we implement an actual fix.