[autoformat] Fix autoformat. (#23941)
For some reason 'dotnet format' doesn't work when:

* The .NET version we're using (in global.json) is .NET 10
* The system .NET is not .NET 10

or the other way around.

_even when the system .NET is executed outside the root repository directory_.

So fix this by installing the .NET version we use on the bot before doing the formatting.

Also update to an autoformat version that supports skipping checking out the source code (because we've already done that).

Backport of #23918.