[CoreFoundation] Improve the handling of data callbacks in CFSocket.OnCallback. (#23955)
Augment CFSocketDataEventArgs to take and store a CFData, and if instantiated
this way, then compute the byte array for the Data property from the CFData
instance on demand instead of when instantiating the CFSocketDataEventArgs.
This way any problems when creating the byte array can be handled in user
code, instead of crashing the process because it occurs in a callback on a
background thread where the app developer can't handle the exception.

Fixes https://github.com/dotnet/macios/issues/18551.