[AudioToolbox] Improve AudioQueue API a little bit. Fixes #11183. (#23988)
In the issue #11183, it becomes the API to allocate and use audio queue buffers is
less than ideal, because the developer has to cast IntPtrs to AudioQueueBuffer pointers.

This is not obvious at all, so create a better API where we return AudioQueueBuffer
pointers directly.

Also improve the API in the AudioQueueBuffer struct to support spans + improve xml docs as well.

And finally add more tests.

Fixes https://github.com/dotnet/macios/issues/11183.