[tests] Attempt to fix a random crash in UIDocumentTest.
Attempt to fix a random crash in UIDocumentTest by using a local variable
(that is later disposed) instead of an instance variable.

The idea is that the instance variable might be freed _during_ the Save method
call, while the local variable won't, because it's implicitly disposed
(because of the using clause).

Hopefully fixes https://github.com/xamarin/maccore/issues/2283.