Revert "[dotnet] Pass --preserve-symbol-paths to the trimmer to ensure trimmed assemblies aren't different between architectures. (#23858)"
This reverts commit dc79f3e89f1e122b2ccd94a9a35ae634d1a4e1ff.

The problem is that we need the assembly to point to the linked version of the
pdb, otherwise the debugger will refuse to load debug info, saying the pdb
doesn't match the dll.

Eventually we might need a way to write just the filename of the pdb in the
assembly, and not a full path, that that's a much more involved fix.

Fixes https://devdiv.visualstudio.com/DevDiv/_workitems/edit/2614232.